requirements.txt
asgiref==3.6.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
Django==4.2.1
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
